analyses list back in the main platform page
