+++ pages/platforms/_cid/index.vue
@@ -2,29 +2,90 @@
   <section>
     <v-layout row justify-space-between>
       <v-btn flat router exact :to="{ path: '/' }"><v-icon left>mdi-arrow-left</v-icon>{{ $t('ui.back') }}</v-btn>
-      <v-btn flat router exact :to="{ name: 'platforms-cid-analyses', params: { cid: $route.params.cid } }">Analyses ({{ analyses.length }}) <v-icon right>mdi-arrow-right</v-icon></v-btn>
     </v-layout>
 
+    <v-dialog v-model="membersDialog" max-width="600">
+      <v-card>
+        <v-card-title class="headline">
+          <span>{{ $t('card.contributors') }}</span>
+          <v-spacer></v-spacer>
+          <v-btn icon @click="membersDialog = false"><v-icon>mdi-close</v-icon></v-btn>
+        </v-card-title>
+
+        <v-list>
+          <v-list-tile v-for="member in card.members" v-bind:key="member.id" avatar :href="'https://trello.com/' + member.username">
+            <v-list-tile-avatar>
+              <img v-if="member.avatarHash" :src="'https://trello-avatars.s3.amazonaws.com/' + member.avatarHash + '/50.png'" alt="avatar">
+              <span v-else class="icon blue-grey lighten-4">{{ member.initials }}</span>
+            </v-list-tile-avatar>
+            <v-list-tile-content>
+              <v-list-tile-title v-text="member.fullName" />
+            </v-list-tile-content>
+          </v-list-tile>
+        </v-list>
+      </v-card>
+    </v-dialog>
+
     <v-card>
-      <v-toolbar class="secondary" dense dark card>
+      <v-toolbar class="secondary" dark dense card>
         <v-toolbar-title>
           {{ card.name }}
         </v-toolbar-title>
 
         <v-spacer></v-spacer>
 
-        <v-tooltip left>
-          <v-btn slot="activator" icon tag="a" flat target="_blank" v-if="card.githubUrl" :href="card.githubUrl"><v-icon>mdi-github-box</v-icon></v-btn>
-          <span>{{ $t('card.github') }}</span>
+        <v-tooltip bottom>
+          <v-btn icon slot="activator" v-if="canEdit" @click="createAnalysis"><v-icon>mdi-plus</v-icon></v-btn>
+          <span>{{ $t('analyses.new') }}</span>
         </v-tooltip>
-        <v-tooltip left>
-          <v-btn slot="activator" icon tag="a" flat target="_blank" v-if="card.homeUrl" :href="card.homeUrl"><v-icon>mdi-home</v-icon></v-btn>
-          <span>{{ $t('card.homepage') }}</span>
-        </v-tooltip>
-        <v-tooltip left>
-          <v-btn slot="activator" icon tag="a" flat target="_blank" v-if="card.url" :href="card.url"><v-icon>mdi-trello</v-icon></v-btn>
-          <span>{{ $t('card.trello') }}</span>
-        </v-tooltip>
+
+        <v-menu>
+          <v-btn slot="activator" icon dark>
+            <v-icon>mdi-dots-vertical</v-icon>
+          </v-btn>
+          <v-list>
+            <v-list-tile avatar v-if="card.members.length > 0" @click="membersDialog = true">
+              <v-list-tile-avatar>
+                <v-icon>mdi-account-multiple</v-icon>
+              </v-list-tile-avatar>
+              <v-list-tile-content>
+                <v-list-tile-title>{{ $t('card.contributors') }}</v-list-tile-title>
+              </v-list-tile-content>
+            </v-list-tile>
+            <v-list-tile avatar v-if="card.githubUrl" :href="card.githubUrl" target="_blank">
+              <v-list-tile-avatar>
+                <v-icon>mdi-github-box</v-icon>
+              </v-list-tile-avatar>
+              <v-list-tile-content>
+                <v-list-tile-title>{{ $t('card.github') }}</v-list-tile-title>
+              </v-list-tile-content>
+            </v-list-tile>
+            <v-list-tile avatar v-if="card.homeUrl" :href="card.homeUrl" target="_blank">
+              <v-list-tile-avatar>
+                <v-icon>mdi-home</v-icon>
+              </v-list-tile-avatar>
+              <v-list-tile-content>
+                <v-list-tile-title>{{ $t('card.homepage') }}</v-list-tile-title>
+              </v-list-tile-content>
+            </v-list-tile>
+            <v-list-tile avatar v-if="card.url" :href="card.url" target="_blank">
+              <v-list-tile-avatar>
+                <v-icon>mdi-trello</v-icon>
+              </v-list-tile-avatar>
+              <v-list-tile-content>
+                <v-list-tile-title>{{ $t('card.trello') }}</v-list-tile-title>
+              </v-list-tile-content>
+            </v-list-tile>
+            <v-list-tile avatar @click="generateTestFile">
+              <v-list-tile-avatar>
+                <v-icon>mdi-upload</v-icon>
+              </v-list-tile-avatar>
+              <v-list-tile-content>
+                <v-list-tile-title>{{ $t('analyses.export') }}</v-list-tile-title>
+              </v-list-tile-content>
+            </v-list-tile>
+          </v-list>
+        </v-menu>
       </v-toolbar>
 
       <v-card-text>
@@ -41,9 +102,9 @@
         </v-layout>
       </v-card-text>
 
-      <v-divider/>
 
       <template v-if="card.humanCertified || card.publisherCertified">
+        <v-divider/>
         <v-subheader>Certifications</v-subheader>
 
         <v-list>
@@ -69,30 +130,39 @@
         <v-divider/>
       </template>
 
+      <v-divider/>
 
-      <v-subheader>{{ $t('card.contributors') }}</v-subheader>
-      <v-list>
-        <v-list-tile v-for="member in card.members" v-bind:key="member.id" avatar :href="'https://trello.com/' + member.username">
-          <v-list-tile-avatar>
-            <img v-if="member.avatarHash" :src="'https://trello-avatars.s3.amazonaws.com/' + member.avatarHash + '/50.png'" alt="avatar">
-            <span v-else class="icon blue-grey lighten-4">{{ member.initials }}</span>
-          </v-list-tile-avatar>
-          <v-list-tile-content>
-            <v-list-tile-title v-text="member.fullName" />
-          </v-list-tile-content>
-        </v-list-tile>
+      <v-list two-line>
+        <draggable v-model="analyses">
+          <AnalysisTile v-for="analysis in analyses" :key="analysis.id" :analysis="analysis" :cardID="card.id"/>
+        </draggable>
       </v-list>
-
     </v-card>
+
   </section>
 </template>
 
 <script>
 import moment from 'moment'
+import AnalysisTile from '~/components/AnalysisTile'
+import draggable from 'vuedraggable'
+import { saveAs } from 'file-saver'
+
+function escapeCSVstring (str) {
+  if (/[";]/.test(str)) {
+    return `"${str.replace(/"/g, '""')}"`
+  } else {
+    return str || ''
+  }
+}
 
 export default {
   name: 'platform',
   transition: 'slide-x-transition',
+  components: {
+    AnalysisTile,
+    draggable
+  },
   async fetch ({ params, store, error }) {
     await store.dispatch('FETCH_TRELLO_LISTS')
 
@@ -110,18 +180,109 @@ export default {
       title: `Platform: ${this.card.name}`
     }
   },
+  data () {
+    return {
+      membersDialog: false
+    }
+  },
   computed: {
     card () {
       return this.$store.state.card
     },
-    analyses () {
-      return this.$store.state.analyses
+    analyses: {
+      get () {
+        return this.$store.state.analyses.sort((a, b) => {
+          return a.order > b.order ? 1 : -1
+        })
+      },
+      async set (list) {
+        if (!this.canEdit) { return }
+
+        try {
+          this.$store.dispatch('REORDER_ANALYSES', { cardID: this.card.id, list })
+        } catch (e) {
+          // eslint-disable-next-line
+          console.error('Reorder failed', e)
+          // TODO: handle error
+        }
+      }
+    },
+    user () {
+      return this.$store.state.user
+    },
+    canEdit () {
+      return this.user && this.user.isAuthorized
     },
     list () {
       return this.$store.state.trelloLists.find(l => this.card.idList === l.id)
     },
     lastActivity () {
       return moment(this.card.lastActivity).locale(this.$i18n.locale).fromNow()
+    }
+  },
+  methods: {
+    async createAnalysis () {
+      this.creating = true
+
+      try {
+        const analysis = await this.$store.dispatch('SAVE_ANALYSIS', { cardID: this.card.id, analysis: {} })
+
+        if (this.card.idMembers.indexOf(this.user.id) === -1) {
+          await this.$store.dispatch('ADD_CARD_MEMBER', {
+            card: this.card,
+            user: this.user
+          })
+        }
+
+        this.$router.push(`/platforms/${this.card.id}/analyses/${analysis.id}/edit`)
+      } catch (e) {
+        // eslint-disable-next-line
+        console.error('Analysis creation failed', e)
+        // TODO: handle error
+      }
+
+      this.creating = false
+    },
+    generateTestFile () {
+      if (!this.analyses) { return }
+
+      const columns = [
+        { title: 'out-unitid', getter (a) { return a.unitid } },
+        { title: 'out-rtype', getter (a) { return a.rtype } },
+        { title: 'out-mime', getter (a) { return a.mime } },
+        { title: 'in-url', getter (a) { return a.url } }
+      ]
+
+      // Add a column for each identifier
+      this.analyses.forEach(analysis => {
+        if (!analysis.identifiers) { return }
+
+        analysis.identifiers.forEach(id => {
+          if (!id.type) { return }
+          if (columns.find(c => c.title === `out-${id.type}`)) { return }
+
+          columns.unshift({
+            title: `out-${id.type}`,
+            getter (a) {
+              if (a.identifiers) {
+                const identifier = a.identifiers.find(i => i.type === id.type)
+                return identifier && identifier.value
+              }
+            }
+          })
+        })
+      })
+
+      const header = columns.map(col => escapeCSVstring(col.title)).join(';')
+
+      const lines = this.analyses.map(analysis => {
+        return columns.map(col => escapeCSVstring(col.getter(analysis))).join(';')
+      }).join('\n')
+
+      const shortName = (/\[([\w\d]+)\]$/.exec(this.card && this.card.name) || [])[1]
+      const fileName = `${shortName || 'test'}.${moment().format('YYYY-MM-DD')}.csv`
+
+      saveAs(new Blob([`${header}\n${lines}`], { type: 'text/csv;charset=utf-8' }), fileName)
     }
   }
 }
